+++ packages/fat-enums/tests/Fixtures/ModelBackedTestEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ArtisanBuild\FatEnums\TestEnums;
+namespace ArtisanBuild\FatEnums\Tests\Fixtures;
 
 use ArtisanBuild\FatEnums\Tests\Fixtures\Models\DummyModel;
 use ArtisanBuild\FatEnums\Traits\DatabaseRecordsEnum;
