+++ packages/fat-enums/tests/Fixtures/UnbackedEnum.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ArtisanBuild\FatEnums\TestEnums;
+namespace ArtisanBuild\FatEnums\Tests\Fixtures;
 
 use ArtisanBuild\FatEnums\Traits\HasBackedEnumMethods;
